requirements.txt
aiogram==3.20.0.post0
APScheduler==3.11.0
pandas==2.2.3
matplotlib==3.10.3
aiohttp==3.12.14
SQLAlchemy==2.0.41
aiosqlite==0.21.0
greenlet==3.2.2
asyncpg==0.30.0
Pillow>=10.0.0
pillow-heif>=0.22.0
python-dotenv>=1.0.0
